Refactor review-related models.
