+++ internals/detect_intent.py
@@ -22,6 +22,7 @@ from framework import permissions
 from framework import users
 from internals import approval_defs
 from internals import models
+from internals import review_models
 
 
 FIELDS_REQUIRING_LGTMS = [
@@ -144,7 +145,7 @@ def is_lgtm_allowed(from_addr, feature, approval_field):
 
 def detect_new_thread(feature_id, approval_field):
   """Return True if there are no previous approval values for this intent."""
-  existing_approvals = models.Approval.get_approvals(
+  existing_approvals = review_models.Approval.get_approvals(
       feature_id=feature_id, field_id=approval_field.field_id)
   return not existing_approvals
 
@@ -281,9 +282,9 @@ class IntentEmailHandler(basehandlers.FlaskHandler):
     if (detect_lgtm(body) and
         is_lgtm_allowed(from_addr, feature, approval_field)):
       logging.info('found LGTM')
-      models.Approval.set_approval(
+      review_models.Approval.set_approval(
           feature_id, approval_field.field_id,
-          models.Approval.APPROVED, from_addr)
+          review_models.Approval.APPROVED, from_addr)
       self.record_lgtm(feature, approval_field, from_addr)
 
     # Case 2: Create a review request for any discussion that does not already
@@ -292,9 +293,9 @@ class IntentEmailHandler(basehandlers.FlaskHandler):
       logging.info('found new thread')
       if approval_field in FIELDS_REQUIRING_LGTMS:
         logging.info('requesting a review')
-        models.Approval.set_approval(
+        review_models.Approval.set_approval(
             feature_id, approval_field.field_id,
-            models.Approval.REVIEW_REQUESTED, from_addr)
+            review_models.Approval.REVIEW_REQUESTED, from_addr)
 
   def record_lgtm(self, feature, approval_field, from_addr):
     """Add from_addr to the old way or recording LGTMs."""
